requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.3
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.4
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
idna==2.9
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-django==3.9.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.25.9
